+++ lib/platformos_check/checks/unreachable_code.rb
@@ -97,13 +97,13 @@ module PlatformosCheck
     def include_node_contains_flow_command?(node)
       return false if node.nil?
 
-      node.nodelist.any? do |node|
-        if INCLUDE_FLOW_COMMAND.include?(node.respond_to?(:tag_name) && node.tag_name)
+      node.nodelist.any? do |n|
+        if INCLUDE_FLOW_COMMAND.include?(n.respond_to?(:tag_name) && n.tag_name)
           true
-        elsif node.respond_to?(:nodelist) && node.nodelist
-          include_node_contains_flow_command?(node)
-        elsif node.respond_to?(:tag_name) && node.tag_name == 'include' && node.template_name_expr.is_a?(String)
-          evaluate_include(node.template_name_expr)
+        elsif n.respond_to?(:nodelist) && n.nodelist
+          include_node_contains_flow_command?(n)
+        elsif n.respond_to?(:tag_name) && n.tag_name == 'include' && n.template_name_expr.is_a?(String)
+          evaluate_include(n.template_name_expr)
         else
           false
         end